fixed error in deleting when finished, add profile api, fixed url of profile images

+++ members/signals.py
@@ -6,5 +6,5 @@ from .models import Class, Attendance
 def delete_attendances_if_class_not_available(sender, instance, **kwargs):
     if instance.pk:
         original_instance = Class.objects.get(pk=instance.pk)
-        if original_instance.is_available_this_week and not instance.is_available_this_week:
+        if original_instance.is_available_this_week and not instance.is_available_this_week and not original_instance.finished and not instance.finished:
             Attendance.objects.filter(class_attended=instance).delete()
